requirements.txt
# Web scraping and API requests
requests==2.28.2

# Asynchronous programming
asyncio==3.4.3
nest-asyncio==1.5.6

# Telegram Bot API
python-telegram-bot==20.0

# Retry logic
tenacity==8.2.1

# Logging and queue management
logging==0.5.1.2

# JSON handling and configuration management
jsonschema==4.18.0

# Miscellaneous
aiohttp==3.8.1
